refactor(esbuild): split up plugin source
